Fix router for nav bar, fix scroll for right column, add Team page
